# WIP: speed up for slow query logs retrieving 