### What problem does this PR solve?

Issue Number: close #20890 

Problem Summary:

The original idea is speed up retrieving slow query logs entry via concurrent read multiple files of slow query log.
But I found problem that GC cost CPU resource far greater than file reading, maybe we should  decrease GC costs first.

#### CPU Profiling

```
Type: cpu
Time: Mar 19, 2021 at 4:41pm (CST)
Duration: 18.61s, Total samples = 30.58s (164.36%)
Showing nodes accounting for 28.74s, 93.98% of 30.58s total
Dropped 196 nodes (cum <= 0.15s)
----------------------------------------------------------+-------------
      flat  flat%   sum%        cum   cum%   calls calls% + context 	 	 
----------------------------------------------------------+-------------
                                            19.27s   100% |   github.com/pingcap/tidb/executor.(*slowQueryRetriever).parseSlowLog.func1
     0.05s  0.16%  0.16%     19.27s 63.02%                | github.com/pingcap/tidb/executor.(*slowQueryRetriever).parseLog
                                            17.92s 92.99% |   github.com/pingcap/tidb/executor.(*slowQueryTuple).convertToDatumRow
                                             0.59s  3.06% |   runtime.newobject
                                             0.52s  2.70% |   github.com/pingcap/tidb/executor.(*slowQueryTuple).setFieldValue
                                             0.07s  0.36% |   runtime.growslice
                                             0.06s  0.31% |   runtime.mallocgc
----------------------------------------------------------+-------------
         0     0%  0.16%     19.27s 63.02%                | github.com/pingcap/tidb/executor.(*slowQueryRetriever).parseSlowLog.func1
                                            19.27s   100% |   github.com/pingcap/tidb/executor.(*slowQueryRetriever).parseLog
----------------------------------------------------------+-------------
                                            17.92s   100% |   github.com/pingcap/tidb/executor.(*slowQueryRetriever).parseLog
     0.83s  2.71%  2.88%     17.92s 58.60%                | github.com/pingcap/tidb/executor.(*slowQueryTuple).convertToDatumRow
                                            11.28s 62.95% |   runtime.growslice
                                             3.39s 18.92% |   runtime.makeslice
                                             2.16s 12.05% |   runtime.typedmemmove
                                             0.14s  0.78% |   runtime.mallocgc
----------------------------------------------------------+-------------
                                             6.59s 43.53% |   runtime.gcBgMarkWorker
                                             4.58s 30.25% |   runtime.gcAssistAlloc
                                             3.63s 23.98% |   runtime.mstart
                                             0.23s  1.52% |   runtime.(*mheap).alloc
                                             0.05s  0.33% |   runtime.bulkBarrierPreWrite
     0.02s 0.065%  2.94%     15.14s 49.51%                | runtime.systemstack
                                             6.69s 44.19% |   runtime.gcBgMarkWorker.func2
                                             4.59s 30.32% |   runtime.gcAssistAlloc.func1
                                             3.65s 24.11% |   runtime.(*mheap).alloc.func1
                                             0.06s   0.4% |   runtime.lock (inline)
----------------------------------------------------------+-------------
                                            11.28s 98.86% |   github.com/pingcap/tidb/executor.(*slowQueryTuple).convertToDatumRow
                                             0.07s  0.61% |   github.com/pingcap/tidb/executor.(*slowQueryRetriever).parseLog
                                             0.05s  0.44% |   github.com/pingcap/tidb/executor.(*slowQueryRetriever).getBatchLog
     0.01s 0.033%  2.98%     11.41s 37.31%                | runtime.growslice
                                             6.72s 58.90% |   runtime.memmove
                                             4.62s 40.49% |   runtime.mallocgc
----------------------------------------------------------+-------------
                                             5.16s 53.92% |   runtime.gcDrain
                                             4.41s 46.08% |   runtime.gcDrainN
     8.67s 28.35% 31.33%      9.57s 31.29%                | runtime.scanobject
                                             0.57s  5.96% |   runtime.heapBits.bits (inline)
                                             0.15s  1.57% |   runtime.findObject
                                             0.14s  1.46% |   runtime.heapBits.next (inline)
----------------------------------------------------------+-------------
                                             4.62s 51.62% |   runtime.growslice
                                             3.52s 39.33% |   runtime.makeslice
                                             0.61s  6.82% |   runtime.newobject
                                             0.14s  1.56% |   github.com/pingcap/tidb/executor.(*slowQueryTuple).convertToDatumRow
                                             0.06s  0.67% |   github.com/pingcap/tidb/executor.(*slowQueryRetriever).parseLog
     0.17s  0.56% 31.88%      8.95s 29.27%                | runtime.mallocgc
                                             4.58s 51.17% |   runtime.gcAssistAlloc
                                             3.76s 42.01% |   runtime.(*mcache).nextFree
                                             0.18s  2.01% |   runtime.heapBitsSetType
                                             0.13s  1.45% |   runtime.memclrNoHeapPointers
----------------------------------------------------------+-------------
                                             6.72s 97.82% |   runtime.growslice
                                             0.07s  1.02% |   runtime.typedmemmove
                                             0.04s  0.58% |   runtime.gcDrain
     6.87s 22.47% 54.35%      6.87s 22.47%                | runtime.memmove

```

#### Mem profiling

```
Type: alloc_objects
Time: Mar 19, 2021 at 4:42pm (CST)
Showing nodes accounting for 6641325, 98.29% of 6756826 total
Dropped 55 nodes (cum <= 33784)
----------------------------------------------------------+-------------
      flat  flat%   sum%        cum   cum%   calls calls% + context 	 	 
----------------------------------------------------------+-------------
                                           4400210   100% |   github.com/pingcap/tidb/executor.(*slowQueryRetriever).parseSlowLog.func1
   1798105 26.61% 26.61%    4400210 65.12%                | github.com/pingcap/tidb/executor.(*slowQueryRetriever).parseLog
                                           2602105 59.14% |   github.com/pingcap/tidb/executor.(*slowQueryTuple).convertToDatumRow
----------------------------------------------------------+-------------
         0     0% 26.61%    4400210 65.12%                | github.com/pingcap/tidb/executor.(*slowQueryRetriever).parseSlowLog.func1
                                           4400210   100% |   github.com/pingcap/tidb/executor.(*slowQueryRetriever).parseLog
----------------------------------------------------------+-------------
                                           2602105   100% |   github.com/pingcap/tidb/executor.(*slowQueryRetriever).parseLog
   1979504 29.30% 55.91%    2602105 38.51%                | github.com/pingcap/tidb/executor.(*slowQueryTuple).convertToDatumRow
                                            622601 23.93% |   github.com/pingcap/tidb/types.NewTimeDatum (inline)
----------------------------------------------------------+-------------
         0     0% 55.91%    2228510 32.98%                | github.com/pingcap/tidb/executor.(*slowQueryRetriever).parseDataForSlowLog
                                           2225989 99.89% |   github.com/pingcap/tidb/executor.(*slowQueryRetriever).parseSlowLog
----------------------------------------------------------+-------------
                                           2225989   100% |   github.com/pingcap/tidb/executor.(*slowQueryRetriever).parseDataForSlowLog
     56956  0.84% 56.75%    2225989 32.94%                | github.com/pingcap/tidb/executor.(*slowQueryRetriever).parseSlowLog
                                           2169033 97.44% |   github.com/pingcap/tidb/executor.(*slowQueryRetriever).getBatchLog
----------------------------------------------------------+-------------
                                           2169033   100% |   github.com/pingcap/tidb/executor.(*slowQueryRetriever).parseSlowLog
     49969  0.74% 57.49%    2169033 32.10%                | github.com/pingcap/tidb/executor.(*slowQueryRetriever).getBatchLog
                                           2119064 97.70% |   github.com/pingcap/tidb/executor.getOneLine
----------------------------------------------------------+-------------
                                           2119064   100% |   github.com/pingcap/tidb/executor.(*slowQueryRetriever).getBatchLog
   2119064 31.36% 88.85%    2119064 31.36%                | github.com/pingcap/tidb/executor.getOneLine
----------------------------------------------------------+-------------
                                            622601   100% |   github.com/pingcap/tidb/types.NewTimeDatum (inline)
    622601  9.21% 98.07%     622601  9.21%                | github.com/pingcap/tidb/types.(*Datum).SetMysqlTime
```

### What is changed and how it works?

Proposal: avoid to alloc ton of objects, decrease GC costs first.

What's Changed:

Add benchmark test(for now)

### Check List

Tests

- Unit test

### Release note

- No